requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.4.0
asyncer==0.0.7
bidict==0.23.1
certifi==2024.8.30
chainlit==1.1.404
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
dataclasses-json==0.5.14
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.110.3
filetype==1.2.0
googleapis-common-protos==1.65.0
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
importlib_metadata==8.4.0
jiter==0.5.0
Lazify==0.4.0
literalai==0.0.607
marshmallow==3.22.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.43.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
packaging==23.2
protobuf==4.25.4
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.4
requests==2.32.3
simple-websocket==1.0.0
sniffio==1.3.1
starlette==0.37.2
syncer==2.0.3
tomli==2.0.1
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
uptrace==1.26.0
urllib3==2.2.2
uvicorn==0.25.0
watchfiles==0.20.0
wrapt==1.16.0
wsproto==1.2.0
zipp==3.20.1
